スポット選択ダイアログのレイアウト調節
